requirements.txt
flask
numpy
pillow
scipy
scikit-learn
joblib
